reorganize files/members
