+++ src/WidgetWorkspace/components/WidgetControlPanel/index.tsx
@@ -1,10 +1,12 @@
 import React from 'react';
 import { observer } from 'mobx-react';
-import { TweakableChildrenInputs } from '../../common/keystone-tweakables/material-ui-controls/TweakableChildrenInputs';
-import { AssetsAccordion } from './AssetsAccordion';
-import { BaseWidgetClass } from '../widget-types/BaseWidgetClass';
-import { useWorkspaceMst } from '../rootStore';
-import { WidgetControlPanelAppBar } from './WidgetControlPanelAppBar';
+import {
+  TweakableChildrenInputs,
+} from '../../../common/keystone-tweakables/material-ui-controls/TweakableChildrenInputs';
+import { AssetsAccordion } from '../AssetsAccordion';
+import { BaseWidgetClass } from '../../widget-types/BaseWidgetClass';
+import { useWorkspaceMst } from '../../rootStore';
+import { WidgetControlPanelAppBar } from './components/WidgetControlPanelAppBar';
 
 export const WidgetControlPanel = observer(() => {
   const workspaceStore = useWorkspaceMst();
