+++ src/WidgetWorkspace/widget-types/SolitaryAssetDefinition.tsx
@@ -2,7 +2,7 @@ import { computed, makeObservable } from 'mobx';
 import ReactDOMServer from 'react-dom/server';
 import React from 'react';
 import { GridPattern } from '../components/ResizableZoomPan/components/GridPattern';
-import { SVGWrapper } from '../components/SVGWrapper';
+import { SVGWrapper } from '../../common/components/SVGWrapper';
 import {
   BaseAssetDefinition, DocumentAreaProps, filePathConstructor, WidgetSVGComponent,
 } from './types';
